+++ CHANGELOG.md
@@ -1,3 +1,51 @@
+# [2.2.0](https://github.com/e-Spirit/fsxa-ui/compare/v2.1.1...v2.2.0) (2021-01-19)
+
+
+### Bug Fixes
+
+* 🐛 Add hover as variant to opacity ([4a91561](https://github.com/e-Spirit/fsxa-ui/commit/4a91561694006bde962f3745134821b552cf8265))
+* 🐛 fix responsive layout of headersection ([#42](https://github.com/e-Spirit/fsxa-ui/issues/42)) ([2391092](https://github.com/e-Spirit/fsxa-ui/commit/23910923e99c26897a082e1bc3fac208e5fc2cc1))
+
+
+### Features
+
+* Slotify Googlemaps section ([1998a39](https://github.com/e-Spirit/fsxa-ui/commit/1998a394e1504045a0b30edea99ab466b6845272))
+
+## [2.1.1](https://github.com/e-Spirit/fsxa-ui/compare/v2.1.0...v2.1.1) (2021-01-19)
+
+
+### Bug Fixes
+
+* 🐛 Optimize FullWidthSliderSection animation ([113e497](https://github.com/e-Spirit/fsxa-ui/commit/113e497275be191c2aafea5c70ddc74e0f638b56))
+
+# [2.1.0](https://github.com/e-Spirit/fsxa-ui/compare/v2.0.4...v2.1.0) (2021-01-19)
+
+
+### Features
+
+* 🎸 Pass slide and slideIndex to click event of FWSSection ([b520429](https://github.com/e-Spirit/fsxa-ui/commit/b5204292fe3e423be9014317df080c05ed3fd5bc))
+
+## [2.0.4](https://github.com/e-Spirit/fsxa-ui/compare/v2.0.3...v2.0.4) (2021-01-18)
+
+
+### Bug Fixes
+
+* 🐛 rename onSlideAnimation to handleSlideAnimation ([b48a322](https://github.com/e-Spirit/fsxa-ui/commit/b48a322cde5ae450a094b19418ae7a17e40347f8))
+
+## [2.0.3](https://github.com/e-Spirit/fsxa-ui/compare/v2.0.2...v2.0.3) (2021-01-18)
+
+
+### Bug Fixes
+
+* 🐛 Remove bottom padding from FullWidthSliderSection ([fe39b51](https://github.com/e-Spirit/fsxa-ui/commit/fe39b514add8677d06161b6cab73bfb41b134671))
+
+## [2.0.2](https://github.com/e-Spirit/fsxa-ui/compare/v2.0.1...v2.0.2) (2021-01-18)
+
+
+### Bug Fixes
+
+* 🐛 Add all width classes to tailwind whitelist ([d932430](https://github.com/e-Spirit/fsxa-ui/commit/d932430b13ecc01f7847a1ca1fb6593ba444b3b1))
+
 ## [2.0.1](https://github.com/e-Spirit/fsxa-ui/compare/v2.0.0...v2.0.1) (2021-01-15)